+++ europeana-eventlistener/src/main/resources/META-INF/jboss-deployment-structure.xml
