+++ keycloak-metrics-spi/src/main/resources/META-INF/services/org.keycloak.services.resource.RealmResourceProviderFactory
@@ -1,0 +1,1 @@
+eu.europeana.keycloak.metrics.UserCountProviderFactory
