+++ keycloak-metrics-spi/src/main/resources/jboss-deployment-structure.xml
