--- keycloak-stats-spi/src/main/java/eu/europeana/keycloak/stats/UserCountProviderFactory.java
@@ -1,38 +1,0 @@
-package eu.europeana.keycloak.stats;
-
-import org.keycloak.Config.Scope;
-import org.keycloak.models.KeycloakSession;
-import org.keycloak.models.KeycloakSessionFactory;
-import org.keycloak.services.resource.RealmResourceProvider;
-import org.keycloak.services.resource.RealmResourceProviderFactory;
-
-/**
- * bogloe
- */
-public class UserCountProviderFactory implements RealmResourceProviderFactory {
-
-    public static final String ID = "stats";
-
-    @Override
-    public String getId() {
-        return ID;
-    }
-
-    @Override
-    public RealmResourceProvider create(KeycloakSession session) {
-        return new UserCountProvider(session);
-    }
-
-    @Override
-    public void init(Scope config) {
-    }
-
-    @Override
-    public void postInit(KeycloakSessionFactory factory) {
-    }
-
-    @Override
-    public void close() {
-    }
-
-}
