--- keycloak-stats-spi/src/main/resources/META-INF/services/org.keycloak.services.resource.RealmResourceProviderFactory
@@ -1,1 +1,0 @@
-eu.europeana.keycloak.stats.UserCountProviderFactory
