--- user-removed-eventlistener/src/main/resources/META-INF/jboss-deployment-structure.xml
@@ -1,12 +1,0 @@
-<jboss-deployment-structure>
-    <deployment>
-        <dependencies>
-            <module name="org.keycloak.keycloak-core"/>
-            <module name="org.keycloak.keycloak-server-spi"/>
-            <module name="org.keycloak.keycloak-server-spi-private"/>
-            <module name="org.keycloak.keycloak-services"/>
-            <module name="org.keycloak.keycloak-saml-core-public"/>
-            <module name="org.jboss.logging"/>
-        </dependencies>
-    </deployment>
-</jboss-deployment-structure>
